old version

+++ src/main/java/com/github/idimabr/raphaspawners/objects/Spawner.java
@@ -4,13 +4,25 @@ import com.github.idimabr.raphaspawners.RaphaSpawners;
 import com.gmail.filoghost.holographicdisplays.api.Hologram;
 import com.gmail.filoghost.holographicdisplays.api.HologramsAPI;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
+import com.mojang.authlib.GameProfile;
+import com.mojang.authlib.properties.Property;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.OfflinePlayer;
 import org.bukkit.block.Block;
 import org.bukkit.block.CreatureSpawner;
 import org.bukkit.entity.EntityType;
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.SkullMeta;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.*;
+import java.util.stream.Collectors;
 
 public class Spawner {
 
@@ -20,6 +32,7 @@ public class Spawner {
     private EntityType type;
     private int quantity;
     private UUID owner;
+    private int delay;
     private HashMap<UUID, List<PermissionType>> members = new HashMap<>();
     private List<SpawnerLog> logs = Lists.newArrayList();
 
@@ -31,13 +44,15 @@ public class Spawner {
         if(block.getType() != Material.MOB_SPAWNER) return;
         CreatureSpawner cs =(CreatureSpawner) block.getState();
         cs.setSpawnedType(type);
-        cs.update(true);
+        cs.update();
 
         this.type = type;
         this.quantity = quantity;
         this.owner = owner;
         this.members = members;
-        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, 2.5, .5));
+        double height = RaphaSpawners.getPlugin().getConfiguration().getDouble("Generator.Hologram.Height");
+        this.delay = RaphaSpawners.getPlugin().getConfiguration().getInt("Generator.Delay.Init");
+        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, height, .5));
         this.logs = logs;
         updateHologram();
     }
@@ -50,13 +65,15 @@ public class Spawner {
         if(block.getType() != Material.MOB_SPAWNER) return;
         CreatureSpawner cs =(CreatureSpawner) block.getState();
         cs.setSpawnedType(type);
-        cs.update(true);
+        cs.update();
 
         this.type = type;
         this.quantity = quantity;
         this.owner = owner;
         this.members = members;
-        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, 2.5, .5));
+        double height = RaphaSpawners.getPlugin().getConfiguration().getDouble("Generator.Hologram.Height");
+        this.delay = RaphaSpawners.getPlugin().getConfiguration().getInt("Generator.Delay.Init");
+        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, height, .5));
         updateHologram();
     }
 
@@ -68,13 +85,15 @@ public class Spawner {
         if(block.getType() != Material.MOB_SPAWNER) return;
         CreatureSpawner cs = (CreatureSpawner) block.getState();
         cs.setSpawnedType(type);
-        cs.update(true);
+        cs.update();
 
         this.type = type;
         this.quantity = 1;
         this.owner = owner;
         this.members = members;
-        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, 2.5, .5));
+        double height = RaphaSpawners.getPlugin().getConfiguration().getDouble("Generator.Hologram.Height");
+        this.delay = RaphaSpawners.getPlugin().getConfiguration().getInt("Generator.Delay.Init");
+        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, height, .5));
         updateHologram();
         RaphaSpawners.getSpawners().put(location, this);
     }
@@ -87,13 +106,15 @@ public class Spawner {
         if(block.getType() != Material.MOB_SPAWNER) return;
         CreatureSpawner cs =(CreatureSpawner) block.getState();
         cs.setSpawnedType(type);
-        cs.update(true);
+        cs.update();
 
         this.type = type;
         this.quantity = quantity;
         this.owner = owner;
         this.members = new HashMap<>();
-        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, 2.5, .5));
+        double height = RaphaSpawners.getPlugin().getConfiguration().getDouble("Generator.Hologram.Height");
+        this.delay = RaphaSpawners.getPlugin().getConfiguration().getInt("Generator.Delay.Init");
+        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, height, .5));
         updateHologram();
         RaphaSpawners.getSpawners().put(location, this);
     }
@@ -106,22 +127,42 @@ public class Spawner {
         if(block.getType() != Material.MOB_SPAWNER) return;
         CreatureSpawner cs =(CreatureSpawner) block.getState();
         cs.setSpawnedType(type);
-        cs.update(true);
+        cs.update();
 
         this.type = type;
         this.quantity = 1;
         this.owner = owner;
         this.members = new HashMap<>();
-        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, 2.5, .5));
+        double height = RaphaSpawners.getPlugin().getConfiguration().getDouble("Generator.Hologram.Height");
+        this.delay = RaphaSpawners.getPlugin().getConfiguration().getInt("Generator.Delay.Init");
+        this.hologram = HologramsAPI.createHologram(RaphaSpawners.getPlugin(), location.clone().add(.5, height, .5));
         updateHologram();
         RaphaSpawners.getSpawners().put(location, this);
     }
 
     public void updateHologram(){
         hologram.clearLines();
-        hologram.appendTextLine("§ex" + quantity + " Gerador" + (quantity > 1 ? "es" : "") + " de " + getEntityName());
-        hologram.appendTextLine("§eProprietário: §f" + Bukkit.getOfflinePlayer(owner).getName());
-        hologram.appendTextLine("§eStatus: §f" + (status ? "§aAtivado" : "§cDesativado"));
+        if(RaphaSpawners.getPlugin().getConfiguration().getBoolean("Generator.MobHead.Enabled")){
+            String URL = RaphaSpawners.getPlugin().getConfigEntities().getString(type.name() + ".HeadURL");
+            hologram.appendItemLine(getCustomSkull(URL));
+        }
+        for (String s : RaphaSpawners.getPlugin().getConfiguration().getStringList("Generator.Hologram.Lines")) {
+            hologram.appendTextLine(
+                    s.replace("&","§")
+                            .replace("%quantity_generators%", quantity+"")
+                            .replace("%owner%", Bukkit.getOfflinePlayer(owner).getName())
+                            .replace("%spawner_type%", getEntityName())
+                            .replace("%spawner_status%", (status ? "§aAtivado" : "§cDesativado"))
+            );
+        }
+    }
+
+    public void hologramDisable(Player player){
+        this.hologram.getVisibilityManager().hideTo(player);
+    }
+
+    public void hologramEnable(Player player){
+        this.hologram.getVisibilityManager().showTo(player);
     }
 
     public List<SpawnerLog> getLogs() {
@@ -166,6 +207,7 @@ public class Spawner {
     }
 
     public void deleteHologram(){
+        if(this.hologram == null) return;
         this.hologram.delete();
     }
 
@@ -202,6 +244,29 @@ public class Spawner {
         this.owner = owner;
     }
 
+    public int getDelay() {
+        return delay;
+    }
+
+    public void setDelay(int delay) {
+        this.delay = delay;
+    }
+
+    public List<Map.Entry<String, Integer>> getTopSpawners(){
+        HashMap<String, Integer> tops = Maps.newHashMap();
+        for (Map.Entry<Location, Spawner> entry : RaphaSpawners.getSpawners().entrySet()) {
+            OfflinePlayer owner = Bukkit.getOfflinePlayer(entry.getValue().getOwner());
+
+            Spawner spawner = entry.getValue();
+
+            tops.put(owner.getName(), tops.getOrDefault(owner.getName(), 0) + spawner.getQuantity());
+        }
+
+        return tops.entrySet()
+                .stream()
+                .sorted(Map.Entry.comparingByValue(Comparator.reverseOrder())).limit(15).collect(Collectors.toList());
+    }
+
     public void addMemberPermission(UUID uuid, PermissionType permission){
         Set<Map.Entry<UUID, List<PermissionType>>> newEntry = members.entrySet();
         for (Map.Entry<UUID, List<PermissionType>> entry : newEntry) {
@@ -233,5 +298,29 @@ public class Spawner {
 
     public void setMembers(HashMap<UUID, List<PermissionType>> members) {
         this.members = members;
+    }
+
+    public ItemStack getCustomSkull(String url) {
+        ItemStack skull = new ItemStack(Material.SKULL_ITEM, 1, (short) 3);
+        if (url == null || url.isEmpty())
+            return skull;
+        SkullMeta skullMeta = (SkullMeta) skull.getItemMeta();
+        GameProfile profile = new GameProfile(UUID.randomUUID(), null);
+        byte[] encodedData = Base64.getEncoder().encode(String.format("{textures:{SKIN:{url:\"%s\"}}}", url).getBytes());
+        profile.getProperties().put("textures", new Property("textures", new String(encodedData)));
+        Field profileField = null;
+        try {
+            profileField = skullMeta.getClass().getDeclaredField("profile");
+        } catch (NoSuchFieldException | SecurityException e) {
+            e.printStackTrace();
+        }
+        profileField.setAccessible(true);
+        try {
+            profileField.set(skullMeta, profile);
+        } catch (IllegalArgumentException | IllegalAccessException e) {
+            e.printStackTrace();
+        }
+        skull.setItemMeta(skullMeta);
+        return skull;
     }
 }
